Move contollers.files.getAll to fs-explorer.list
