--- server/controllers/index.js
@@ -1,4 +1,0 @@
-let files = require('./files');
-let exports = {};
-exports.files = files;
-module.exports = exports;
