+++ server/fs-explorer/index.js
@@ -1,7 +1,48 @@
 let path = require('path'),
+  fs = require('fs'),
+  os = require('os'),
   recursiveReadDir = require('recursive-readdir');
 
+const defaultExcludePattern = {
+  HIDDEN_FILES: /^\..+$/
+};
 const SUPPORTED_MEDIA = ['.mkv', '.mp4'];
+
+let list = (dirPath, excludePattern, includeHidden) => {
+  excludePattern = excludePattern || includeHidden ? null : defaultExcludePattern.HIDDEN_FILES;
+
+  let currentDir = dirPath || os.homedir();
+  let data = [];
+  return new Promise((resolve, reject) => {
+    fs.readdir(currentDir, (err, files) => {
+      if (err) {
+        return reject(err);
+      }
+
+      files
+        .forEach((file) => {
+          if (excludePattern && excludePattern.test(file)) return;
+          try {
+            let filePath = path.join(currentDir, file);
+            let isDirectory = fs.statSync(filePath).isDirectory();
+
+            let fileInfo = {
+              name: file,
+              path: filePath,
+              ext: isDirectory ? '' : path.extname(file),
+              isDirectory: isDirectory
+            };
+
+            data.push(fileInfo);
+          } catch (e) {
+            return reject(e);
+          }
+        });
+
+      return resolve(data);
+    });
+  });
+};
 
 let scanForMedia = function (root, callback) {
   let ignoreFunc = function (file, stats) {
@@ -16,5 +57,6 @@ let scanForMedia = function (root, callback) {
 };
 
 module.exports = {
+  list,
   scanForMedia
 };
