+++ server/routes/files.js
@@ -2,7 +2,6 @@ let express = require('express');
 let async = require('async');
 
 let router = express.Router();
-let controllers = require('../controllers');
 const statusCodes = require('http-status-codes');
 let fileExplorer = require('../fs-explorer');
 let mediaProcessor = require('../media');
@@ -11,7 +10,7 @@ let mediaProcessor = require('../media');
  * Returns files and folders inside a given folder path or home directory as default.
  */
 router.get('/files', function (req, res) {
-  controllers.files.getAll(req.query.path)
+  fileExplorer.list(req.query.path)
     .then((files) => {
       return res.status(statusCodes.OK).json(files);
     })
